requirements.txt
discord.py==2.7.1
pandas==3.0.1
